Makes filter bar hideable.
Unlike the DC codebase, we're just showing the filter toggle at all
sizes (it's arguably useful for big displays too).

This commit adds a new Backbone view wrapper around the `#map-controls`
and captures a click event to hide-show the sidebar via CSS. It also
makes the mapview full-width under the sidebar to accommodate the closed
state. This is a backbone-ish expansion of the ad-hoc
[reach-across-the-view-hierarchy][1] situation in the DC codebase.

[1]: https://github.com/cityenergyproject/dc/blob/8014c00a84c28e0e9fef10033688ce4be94e2abe/src/app/views/map/map.js#L29